Refactor eth_call spec
