+++ packages/relay/package.json
@@ -33,7 +33,11 @@
     "test:eth-get-transaction-by-block-hash-and-index": "nyc ts-mocha --recursive './tests/**/*.spec.ts' './tests/**/**/*.spec.ts' -g '@ethGetTransactionByBlockHashAndIndex' --exit",
     "test:eth-get-balance": "nyc ts-mocha --recursive './tests/**/*.spec.ts' './tests/**/**/*.spec.ts' -g '@ethGetBalance' --exit",
     "test:eth-get-code": "nyc ts-mocha --recursive './tests/**/*.spec.ts' './tests/**/**/*.spec.ts' -g '@ethGetCode' --exit",
-    "test:eth-fee-history": "nyc ts-mocha --recursive './tests/**/*.spec.ts' './tests/**/**/*.spec.ts' -g '@ethFeeHistory' --exit"
+    "test:eth-fee-history": "nyc ts-mocha --recursive './tests/**/*.spec.ts' './tests/**/**/*.spec.ts' -g '@ethFeeHistory' --exit",
+    "test:eth-estimate-gas": "nyc ts-mocha --recursive './tests/**/*.spec.ts' './tests/**/**/*.spec.ts' -g '@ethEstimateGas' --exit",
+    "test:eth-gas-price": "nyc ts-mocha --recursive './tests/**/*.spec.ts' './tests/**/**/*.spec.ts' -g '@ethGasPrice' --exit",
+    "test:eth-call": "nyc ts-mocha --recursive './tests/**/*.spec.ts' './tests/**/**/*.spec.ts' -g '@ethCall' --exit",
+    "test:eth-send-raw-transaction": "nyc ts-mocha --recursive './tests/**/*.spec.ts' './tests/**/**/*.spec.ts' -g '@ethSendRawTransaction' --exit"
   },
   "dependencies": {
     "@ethersproject/asm": "^5.7.0",
