+++ packages/relay/tests/lib/eth.spec.ts
@@ -70,7 +70,6 @@ import { JsonRpcError } from '../../dist';
 import { hashNumber, numberTo0x, nullableNumberTo0x, toHash32 } from '../../dist/formatters';
 import * as _ from 'lodash';
 import { CacheService } from '../../src/lib/services/cacheService/cacheService';
-import { TINYBAR_TO_WEIBAR_COEF_BIGINT } from './eth/config';
 
 chai.use(chaiAsPromised);
 
@@ -541,1427 +540,6 @@ describe('Eth calls using MirrorNode', async function () {
 
   this.beforeEach(() => {
     restMock.onGet('network/fees').reply(200, defaultNetworkFees);
-  });
-
-  it('eth_estimateGas with transaction.data null does not fail', async function () {
-    const callData = {
-      from: '0x05fba803be258049a27b820088bab1cad2058871',
-      value: '0x0',
-      gasPrice: '0x0',
-      data: null,
-    };
-    web3Mock
-      .onPost('contracts/call', { ...callData, estimate: true })
-      .reply(501, { errorMessage: '', statusCode: 501 });
-
-    const gas = await ethImpl.estimateGas(callData, null);
-    expect(gas).to.equal(numberTo0x(constants.TX_DEFAULT_GAS_DEFAULT));
-  });
-
-  it('eth_estimateGas to mirror node for contract call returns 501', async function () {
-    const callData = {
-      data: '0x608060405234801561001057600080fd5b506040516107893803806107898339818101604052810190610032919061015a565b806000908051906020019061004892919061004f565b50506102f6565b82805461005b90610224565b90600052602060002090601f01602090048101928261007d57600085556100c4565b82601f1061009657805160ff19168380011785556100c4565b828001600101855582156100c4579182015b828111156100c35782518255916020019190600101906100a8565b5b5090506100d191906100d5565b5090565b5b808211156100ee5760008160009055506001016100d6565b5090565b6000610105610100846101c0565b61019b565b90508281526020810184848401111561011d57600080fd5b6101288482856101f1565b509392505050565b600082601f83011261014157600080fd5b81516101518482602086016100f2565b91505092915050565b60006020828403121561016c57600080fd5b600082015167ffffffffffffffff81111561018657600080fd5b61019284828501610130565b91505092915050565b60006101a56101b6565b90506101b18282610256565b919050565b6000604051905090565b600067ffffffffffffffff8211156101db576101da6102b6565b5b6101e4826102e5565b9050602081019050919050565b60005b8381101561020f5780820151818401526020810190506101f4565b8381111561021e576000848401525b50505050565b6000600282049050600182168061023c57607f821691505b602082108114156102505761024f610287565b5b50919050565b61025f826102e5565b810181811067ffffffffffffffff8211171561027e5761027d6102b6565b5b80604052505050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052602260045260246000fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b6000601f19601f8301169050919050565b610484806103056000396000f3fe608060405234801561001057600080fd5b50600436106100365760003560e01c8063a41368621461003b578063cfae321714610057575b600080fd5b6100556004803603810190610050919061022c565b610075565b005b61005f61008f565b60405161006c91906102a6565b60405180910390f35b806000908051906020019061008b929190610121565b5050565b60606000805461009e9061037c565b80601f01602080910402602001604051908101604052809291908181526020018280546100ca9061037c565b80156101175780601f106100ec57610100808354040283529160200191610117565b820191906000526020600020905b8154815290600101906020018083116100fa57829003601f168201915b5050505050905090565b82805461012d9061037c565b90600052602060002090601f01602090048101928261014f5760008555610196565b82601f1061016857805160ff1916838001178555610196565b82800160010185558215610196579182015b8281111561019557825182559160200191906001019061017a565b5b5090506101a391906101a7565b5090565b5b808211156101c05760008160009055506001016101a8565b5090565b60006101d76101d2846102ed565b6102c8565b9050828152602081018484840111156101ef57600080fd5b6101fa84828561033a565b509392505050565b600082601f83011261021357600080fd5b81356102238482602086016101c4565b91505092915050565b60006020828403121561023e57600080fd5b600082013567ffffffffffffffff81111561025857600080fd5b61026484828501610202565b91505092915050565b60006102788261031e565b6102828185610329565b9350610292818560208601610349565b61029b8161043d565b840191505092915050565b600060208201905081810360008301526102c0818461026d565b905092915050565b60006102d26102e3565b90506102de82826103ae565b919050565b6000604051905090565b600067ffffffffffffffff8211156103085761030761040e565b5b6103118261043d565b9050602081019050919050565b600081519050919050565b600082825260208201905092915050565b82818337600083830152505050565b60005b8381101561036757808201518184015260208101905061034c565b83811115610376576000848401525b50505050565b6000600282049050600182168061039457607f821691505b602082108114156103a8576103a76103df565b5b50919050565b6103b78261043d565b810181811067ffffffffffffffff821117156103d6576103d561040e565b5b80604052505050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052602260045260246000fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b6000601f19601f830116905091905056fea264697066735822122070d157c4efbb3fba4a1bde43cbba5b92b69f2fc455a650c0dfb61e9ed3d4bd6364736f6c634300080400330000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000000b696e697469616c5f6d7367000000000000000000000000000000000000000000',
-      from: '0x81cb089c285e5ee3a7353704fb114955037443af',
-    };
-    web3Mock
-      .onPost('contracts/call', { ...callData, estimate: true })
-      .reply(501, { errorMessage: '', statusCode: 501 });
-
-    const gas = await ethImpl.estimateGas(callData, null);
-    expect(gas).to.equal(numberTo0x(constants.TX_DEFAULT_GAS_DEFAULT));
-  });
-
-  it('eth_estimateGas contract call returns workaround response from mirror-node', async function () {
-    const callData = {
-      data: '0x608060405234801561001057600080fd5b506040516107893803806107898339818101604052810190610032919061015a565b806000908051906020019061004892919061004f565b50506102f6565b82805461005b90610224565b90600052602060002090601f01602090048101928261007d57600085556100c4565b82601f1061009657805160ff19168380011785556100c4565b828001600101855582156100c4579182015b828111156100c35782518255916020019190600101906100a8565b5b5090506100d191906100d5565b5090565b5b808211156100ee5760008160009055506001016100d6565b5090565b6000610105610100846101c0565b61019b565b90508281526020810184848401111561011d57600080fd5b6101288482856101f1565b509392505050565b600082601f83011261014157600080fd5b81516101518482602086016100f2565b91505092915050565b60006020828403121561016c57600080fd5b600082015167ffffffffffffffff81111561018657600080fd5b61019284828501610130565b91505092915050565b60006101a56101b6565b90506101b18282610256565b919050565b6000604051905090565b600067ffffffffffffffff8211156101db576101da6102b6565b5b6101e4826102e5565b9050602081019050919050565b60005b8381101561020f5780820151818401526020810190506101f4565b8381111561021e576000848401525b50505050565b6000600282049050600182168061023c57607f821691505b602082108114156102505761024f610287565b5b50919050565b61025f826102e5565b810181811067ffffffffffffffff8211171561027e5761027d6102b6565b5b80604052505050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052602260045260246000fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b6000601f19601f8301169050919050565b610484806103056000396000f3fe608060405234801561001057600080fd5b50600436106100365760003560e01c8063a41368621461003b578063cfae321714610057575b600080fd5b6100556004803603810190610050919061022c565b610075565b005b61005f61008f565b60405161006c91906102a6565b60405180910390f35b806000908051906020019061008b929190610121565b5050565b60606000805461009e9061037c565b80601f01602080910402602001604051908101604052809291908181526020018280546100ca9061037c565b80156101175780601f106100ec57610100808354040283529160200191610117565b820191906000526020600020905b8154815290600101906020018083116100fa57829003601f168201915b5050505050905090565b82805461012d9061037c565b90600052602060002090601f01602090048101928261014f5760008555610196565b82601f1061016857805160ff1916838001178555610196565b82800160010185558215610196579182015b8281111561019557825182559160200191906001019061017a565b5b5090506101a391906101a7565b5090565b5b808211156101c05760008160009055506001016101a8565b5090565b60006101d76101d2846102ed565b6102c8565b9050828152602081018484840111156101ef57600080fd5b6101fa84828561033a565b509392505050565b600082601f83011261021357600080fd5b81356102238482602086016101c4565b91505092915050565b60006020828403121561023e57600080fd5b600082013567ffffffffffffffff81111561025857600080fd5b61026484828501610202565b91505092915050565b60006102788261031e565b6102828185610329565b9350610292818560208601610349565b61029b8161043d565b840191505092915050565b600060208201905081810360008301526102c0818461026d565b905092915050565b60006102d26102e3565b90506102de82826103ae565b919050565b6000604051905090565b600067ffffffffffffffff8211156103085761030761040e565b5b6103118261043d565b9050602081019050919050565b600081519050919050565b600082825260208201905092915050565b82818337600083830152505050565b60005b8381101561036757808201518184015260208101905061034c565b83811115610376576000848401525b50505050565b6000600282049050600182168061039457607f821691505b602082108114156103a8576103a76103df565b5b50919050565b6103b78261043d565b810181811067ffffffffffffffff821117156103d6576103d561040e565b5b80604052505050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052602260045260246000fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b6000601f19601f830116905091905056fea264697066735822122070d157c4efbb3fba4a1bde43cbba5b92b69f2fc455a650c0dfb61e9ed3d4bd6364736f6c634300080400330000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000000b696e697469616c5f6d7367000000000000000000000000000000000000000000',
-      from: '0x81cb089c285e5ee3a7353704fb114955037443af',
-    };
-    web3Mock.onPost('contracts/call', { ...callData, estimate: true }).reply(200, { result: `0x61A80` });
-
-    const gas = await ethImpl.estimateGas(callData, null);
-    expect(gas.toLowerCase()).to.equal(numberTo0x(constants.TX_DEFAULT_GAS_DEFAULT).toLowerCase());
-  });
-
-  it('eth_estimateGas contract call returns default', async function () {
-    const gas = await ethImpl.estimateGas({ data: '0x01' }, null);
-    expect(gas).to.equal(numberTo0x(constants.TX_DEFAULT_GAS_DEFAULT));
-  });
-
-  it('eth_estimateGas to mirror node for transfer returns 501', async function () {
-    const callData = {
-      data: '0x',
-      from: '0x81cb089c285e5ee3a7353704fb114955037443af',
-      to: receiverAddress,
-      value: '0x2540BE400',
-    };
-    web3Mock
-      .onPost('contracts/call', { ...callData, estimate: true })
-      .reply(501, { errorMessage: '', statusCode: 501 });
-
-    restMock.onGet(`accounts/${receiverAddress}${noTransactions}`).reply(200, { address: receiverAddress });
-
-    const gas = await ethImpl.estimateGas(callData, null);
-    expect(gas).to.equal(numberTo0x(constants.TX_BASE_COST));
-  });
-
-  it('eth_estimateGas to mirror node for transfer without value returns 501', async function () {
-    const callData = {
-      data: '0x',
-      from: '0x81cb089c285e5ee3a7353704fb114955037443af',
-      to: receiverAddress,
-    };
-    web3Mock
-      .onPost('contracts/call', { ...callData, estimate: true })
-      .reply(501, { errorMessage: '', statusCode: 501 });
-    restMock.onGet(`accounts/${receiverAddress}${noTransactions}`).reply(200, { address: receiverAddress });
-
-    const result = await ethImpl.estimateGas(callData, null);
-    expect(result).to.not.be.null;
-    expect(result.code).to.eq(-32602);
-    expect(result.name).to.eq('Invalid parameter');
-  });
-
-  it('eth_estimateGas transfer to existing account', async function () {
-    restMock.onGet(`accounts/${receiverAddress}${noTransactions}`).reply(200, { address: receiverAddress });
-
-    const gas = await ethImpl.estimateGas(
-      {
-        to: receiverAddress,
-        value: 100_000_000_000,
-      },
-      null,
-    );
-    expect(gas).to.equal(EthImpl.gasTxBaseCost);
-  });
-
-  it('eth_estimateGas transfer to existing cached account', async function () {
-    restMock.onGet(`accounts/${receiverAddress}${noTransactions}`).reply(200, { address: receiverAddress });
-
-    const gasBeforeCache = await ethImpl.estimateGas(
-      {
-        to: receiverAddress,
-        value: 100_000_000_000,
-      },
-      null,
-    );
-
-    restMock.onGet(`accounts/${receiverAddress}${noTransactions}`).reply(404);
-    const gasAfterCache = await ethImpl.estimateGas(
-      {
-        to: receiverAddress,
-        value: 100_000_000_000,
-      },
-      null,
-    );
-
-    expect(gasBeforeCache).to.equal(EthImpl.gasTxBaseCost);
-    expect(gasAfterCache).to.equal(EthImpl.gasTxBaseCost);
-  });
-
-  it('eth_estimateGas transfer to non existing account', async function () {
-    restMock.onGet(`accounts/${receiverAddress}${noTransactions}`).reply(404);
-
-    const hollowAccountGasCreation = await ethImpl.estimateGas(
-      {
-        to: receiverAddress,
-        value: 100_000_000_000,
-      },
-      null,
-    );
-
-    expect(hollowAccountGasCreation).to.equal(EthImpl.gasTxHollowAccountCreation);
-  });
-
-  it('eth_estimateGas transfer with 0 value', async function () {
-    const result = await ethImpl.estimateGas(
-      {
-        to: receiverAddress,
-        value: 0,
-      },
-      null,
-    );
-
-    expect(result).to.exist;
-    expect(result.code).to.equal(-32602);
-    expect(result.name).to.equal('Invalid parameter');
-    expect(result.message).to.equal(
-      `Invalid parameter 0: Invalid 'value' field in transaction param. Value must be greater than 0`,
-    );
-  });
-
-  it('eth_estimateGas transfer with invalid value', async function () {
-    const result = await ethImpl.estimateGas(
-      {
-        to: receiverAddress,
-        value: null,
-      },
-      null,
-    );
-
-    expect(result).to.exist;
-    expect(result.code).to.equal(-32602);
-    expect(result.name).to.equal('Invalid parameter');
-    expect(result.message).to.equal(
-      `Invalid parameter 0: Invalid 'value' field in transaction param. Value must be greater than 0`,
-    );
-  });
-
-  it('eth_estimateGas empty call returns transfer cost', async function () {
-    restMock.onGet(`accounts/undefined${noTransactions}`).reply(404);
-    const gas = await ethImpl.estimateGas({}, null);
-    expect(gas).to.equal(numberTo0x(constants.TX_DEFAULT_GAS_DEFAULT));
-  });
-
-  it('eth_estimateGas empty call returns transfer cost with overridden default gas', async function () {
-    const defaultGasOverride = constants.TX_DEFAULT_GAS_DEFAULT + 1;
-    process.env.TX_DEFAULT_GAS = defaultGasOverride.toString();
-    const ethImplOverridden = new EthImpl(sdkClientStub, mirrorNodeInstance, logger, '0x12a', registry, cacheService);
-    restMock.onGet(`accounts/undefined${noTransactions}`).reply(404);
-    const gas = await ethImplOverridden.estimateGas({}, null);
-    expect(gas).to.equal(numberTo0x(defaultGasOverride));
-  });
-
-  it('eth_estimateGas empty input transfer cost', async function () {
-    restMock.onGet(`accounts/undefined${noTransactions}`).reply(404);
-    const gas = await ethImpl.estimateGas({ data: '' }, null);
-    expect(gas).to.equal(numberTo0x(constants.TX_DEFAULT_GAS_DEFAULT));
-  });
-
-  it('eth_estimateGas empty input transfer cost with overridden default gas', async function () {
-    const defaultGasOverride = constants.TX_DEFAULT_GAS_DEFAULT + 1;
-    process.env.TX_DEFAULT_GAS = defaultGasOverride.toString();
-    const ethImplOverridden = new EthImpl(sdkClientStub, mirrorNodeInstance, logger, '0x12a', registry, cacheService);
-    restMock.onGet(`accounts/undefined${noTransactions}`).reply(404);
-    const gas = await ethImplOverridden.estimateGas({ data: '' }, null);
-    expect(gas).to.equal(numberTo0x(defaultGasOverride));
-  });
-
-  it('eth_estimateGas zero input returns transfer cost', async function () {
-    restMock.onGet(`accounts/undefined${noTransactions}`).reply(404);
-    const gas = await ethImpl.estimateGas({ data: '0x' }, null);
-    expect(gas).to.equal(numberTo0x(constants.TX_DEFAULT_GAS_DEFAULT));
-  });
-
-  it('eth_estimateGas zero input returns transfer cost with overridden default gas', async function () {
-    const defaultGasOverride = constants.TX_DEFAULT_GAS_DEFAULT + 1;
-    process.env.TX_DEFAULT_GAS = defaultGasOverride.toString();
-    const ethImplOverridden = new EthImpl(sdkClientStub, mirrorNodeInstance, logger, '0x12a', registry, cacheService);
-    restMock.onGet(`accounts/undefined${noTransactions}`).reply(404);
-    const gas = await ethImplOverridden.estimateGas({ data: '0x' }, null);
-    expect(gas).to.equal(numberTo0x(defaultGasOverride));
-  });
-
-  it('eth_estimateGas with contract revert and message does not equal executionReverted', async function () {
-    const transaction = {
-      from: '0x05fba803be258049a27b820088bab1cad2058871',
-      data: '0x60806040523480156200001157600080fd5b50604051620019f4380380620019f48339818101604052810190620000379190620001fa565b818181600390816200004a9190620004ca565b5080600490816200005c9190620004ca565b5050505050620005b1565b6000604051905090565b600080fd5b600080fd5b600080fd5b600080fd5b6000601f19601f8301169050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b620000d08262000085565b810181811067ffffffffffffffff82111715620000f257620000f162000096565b5b80604052505',
-    };
-    const id = uuid();
-    web3Mock.onPost('contracts/call', { ...transaction, estimate: true }).reply(400, {
-      _status: {
-        messages: [
-          {
-            message: 'data field invalid hexadecimal string',
-            detail: '',
-            data: '',
-          },
-        ],
-      },
-    });
-
-    const result: any = await ethImpl.estimateGas(transaction, id);
-
-    expect(result).to.equal(numberTo0x(constants.TX_DEFAULT_GAS_DEFAULT));
-  });
-
-  it('eth_estimateGas with contract revert and message does not equal executionReverted and ESTIMATE_GAS_THROWS is set to false', async function () {
-    const estimateGasThrows = process.env.ESTIMATE_GAS_THROWS;
-    process.env.ESTIMATE_GAS_THROWS = 'false';
-    const transaction = {
-      from: '0x05fba803be258049a27b820088bab1cad2058871',
-      data: '0x60806040523480156200001157600080fd5b50604051620019f4380380620019f48339818101604052810190620000379190620001fa565b818181600390816200004a9190620004ca565b5080600490816200005c9190620004ca565b5050505050620005b1565b6000604051905090565b600080fd5b600080fd5b600080fd5b600080fd5b6000601f19601f8301169050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b620000d08262000085565b810181811067ffffffffffffffff82111715620000f257620000f162000096565b5b80604052505',
-    };
-    const id = uuid();
-    web3Mock.onPost('contracts/call', { ...transaction, estimate: true }).reply(400, {
-      _status: {
-        messages: [
-          {
-            message: 'data field invalid hexadecimal string',
-            detail: '',
-            data: '',
-          },
-        ],
-      },
-    });
-
-    const result: any = await ethImpl.estimateGas(transaction, id);
-
-    expect(result).to.equal(numberTo0x(constants.TX_DEFAULT_GAS_DEFAULT));
-    process.env.ESTIMATE_GAS_THROWS = estimateGasThrows;
-  });
-
-  it('eth_estimateGas with contract revert and message equals "execution reverted: Invalid number of recipients"', async function () {
-    const transaction = {
-      from: '0x05fba803be258049a27b820088bab1cad2058871',
-      data: '0x60806040523480156200001157600080fd5b50604051620019f4380380620019f48339818101604052810190620000379190620001fa565b818181600390816200004a9190620004ca565b5080600490816200005c9190620004ca565b5050505050620005b1565b6000604051905090565b600080fd5b600080fd5b600080fd5b600080fd5b6000601f19601f8301169050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b620000d08262000085565b810181811067ffffffffffffffff82111715620000f257620000f162000096565b5b80604052505',
-    };
-    const id = uuid();
-    web3Mock.onPost('contracts/call', { ...transaction, estimate: true }).reply(400, {
-      _status: {
-        messages: [
-          {
-            data: '0x08c379a00000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000001c496e76616c6964206e756d626572206f6620726563697069656e747300000000',
-            detail: 'Invalid number of recipients',
-            message: 'CONTRACT_REVERT_EXECUTED',
-          },
-        ],
-      },
-    });
-
-    const result: any = await ethImpl.estimateGas(transaction, id);
-
-    expect(result.data).to.equal(
-      '0x08c379a00000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000001c496e76616c6964206e756d626572206f6620726563697069656e747300000000',
-    );
-    expect(result.message).to.equal('execution reverted: Invalid number of recipients');
-  });
-
-  it('eth_estimateGas handles a 501 unimplemented response from the mirror node correctly by returning default gas', async function () {
-    const transaction = {
-      from: '0x05fba803be258049a27b820088bab1cad2058871',
-      data: '0x60806040523480156200001157600080fd5b50604051620019f4380380620019f48339818101604052810190620000379190620001fa565b818181600390816200004a9190620004ca565b5080600490816200005c9190620004ca565b5050505050620005b1565b6000604051905090565b600080fd5b600080fd5b600080fd5b600080fd5b6000601f19601f8301169050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b620000d08262000085565b810181811067ffffffffffffffff82111715620000f257620000f162000096565b5b80604052505',
-    };
-    const id = uuid();
-    web3Mock.onPost('contracts/call', { ...transaction, estimate: true }).reply(501, {
-      _status: {
-        messages: [
-          {
-            message: 'Auto account creation is not supported.',
-            detail: '',
-            data: '',
-          },
-        ],
-      },
-    });
-
-    const result: any = await ethImpl.estimateGas(transaction, id);
-    expect(result).to.equal(numberTo0x(constants.TX_DEFAULT_GAS_DEFAULT));
-  });
-
-  it('should perform estimateGas precheck', async function () {
-    const transaction = {
-      from: '0x05fba803be258049a27b820088bab1cad2058871',
-      data: '0x',
-      value: '0xA186B8E9800',
-      gasPrice: '0xF4240',
-      gas: '0xd97010',
-    };
-
-    ethImpl.contractCallFormat(transaction);
-    expect(transaction.value).to.eq(1110);
-    expect(transaction.gasPrice).to.eq(1000000);
-    expect(transaction.gas).to.eq(14250000);
-  });
-
-  describe('eth_gasPrice', async function () {
-    it('eth_gasPrice', async function () {
-      restMock.onGet(`network/fees`).reply(200, defaultNetworkFees);
-
-      const weiBars = await ethImpl.gasPrice();
-      const expectedWeiBars = defaultNetworkFees.fees[2].gas * constants.TINYBAR_TO_WEIBAR_COEF;
-      expect(weiBars).to.equal(numberTo0x(expectedWeiBars));
-    });
-
-    it('eth_gasPrice with cached value', async function () {
-      restMock.onGet(`network/fees`).reply(200, defaultNetworkFees);
-
-      const firstGasResult = await ethImpl.gasPrice();
-
-      const modifiedNetworkFees = { ...defaultNetworkFees };
-      modifiedNetworkFees.fees[2].gas = defaultNetworkFees.fees[2].gas * 100;
-
-      restMock.onGet(`network/fees`).reply(200, modifiedNetworkFees);
-
-      const secondGasResult = await ethImpl.gasPrice();
-
-      expect(firstGasResult).to.equal(secondGasResult);
-    });
-
-    it('eth_gasPrice with no EthereumTransaction gas returned', async function () {
-      // deep copy defaultNetworkFees to avoid mutating the original object
-      const partialNetworkFees = JSON.parse(JSON.stringify(defaultNetworkFees));
-      partialNetworkFees.fees.splice(2);
-
-      restMock.onGet(`network/fees`).reply(200, partialNetworkFees);
-
-      await RelayAssertions.assertRejection(predefined.COULD_NOT_ESTIMATE_GAS_PRICE, ethImpl.gasPrice, true, ethImpl);
-    });
-
-    it('eth_gasPrice with mirror node return network fees found', async function () {
-      restMock.onGet(`network/fees`).reply(404, {
-        _status: {
-          messages: [
-            {
-              message: 'Not found',
-            },
-          ],
-        },
-      });
-
-      const fauxGasTinyBars = 35_000;
-      const fauxGasWeiBarHex = '0x13e52b9abe000';
-      sdkClientStub.getTinyBarGasFee.returns(fauxGasTinyBars);
-
-      const gas = await ethImpl.gasPrice();
-      expect(gas).to.equal(fauxGasWeiBarHex);
-    });
-
-    it('eth_gasPrice with no network fees records found', async function () {
-      restMock.onGet(`network/fees`).reply(404, {
-        _status: {
-          messages: [
-            {
-              message: 'Not found',
-            },
-          ],
-        },
-      });
-
-      await RelayAssertions.assertRejection(predefined.COULD_NOT_ESTIMATE_GAS_PRICE, ethImpl.gasPrice, true, ethImpl);
-    });
-  });
-
-  describe('eth_call precheck failures', async function () {
-    let callConsensusNodeSpy: sinon.SinonSpy;
-    let callMirrorNodeSpy: sinon.SinonSpy;
-    let sandbox: sinon.SinonSandbox;
-
-    beforeEach(() => {
-      sandbox = sinon.createSandbox();
-      callConsensusNodeSpy = sandbox.spy(ethImpl, 'callConsensusNode');
-      callMirrorNodeSpy = sandbox.spy(ethImpl, 'callMirrorNode');
-    });
-
-    afterEach(() => {
-      sandbox.restore();
-    });
-
-    it('eth_call with missing `to` field', async function () {
-      await ethCallFailing(
-        ethImpl,
-        {
-          from: contractAddress1,
-          data: contractCallData,
-          gas: maxGasLimitHex,
-        },
-        'latest',
-        (error) => {
-          expect(error.message).to.equal(`Invalid Contract Address: ${undefined}.`);
-        },
-      );
-    });
-
-    it('eth_call with incorrect `to` field length', async function () {
-      await ethCallFailing(
-        ethImpl,
-        {
-          from: contractAddress1,
-          to: EthImpl.zeroHex,
-          data: contractCallData,
-          gas: maxGasLimitHex,
-        },
-        'latest',
-        (error) => {
-          expect(error.message).to.equal(
-            `Invalid Contract Address: ${EthImpl.zeroHex}. Expected length of 42 chars but was 3.`,
-          );
-        },
-      );
-    });
-
-    it('should execute "eth_call" against mirror node with a false ETH_CALL_DEFAULT_TO_CONSENSUS_NODE', async function () {
-      const initialEthCallConesneusFF = process.env.ETH_CALL_DEFAULT_TO_CONSENSUS_NODE;
-
-      process.env.ETH_CALL_DEFAULT_TO_CONSENSUS_NODE = 'false';
-      restMock.onGet(`contracts/${defaultCallData.from}`).reply(404);
-      restMock.onGet(`accounts/${defaultCallData.from}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: defaultCallData.from,
-      });
-      restMock.onGet(`contracts/${defaultCallData.to}`).reply(200, defaultContract);
-      await ethImpl.call({ ...defaultCallData, gas: `0x${defaultCallData.gas.toString(16)}` }, 'latest');
-
-      assert(callMirrorNodeSpy.calledOnce);
-      process.env.ETH_CALL_DEFAULT_TO_CONSENSUS_NODE = initialEthCallConesneusFF;
-    });
-
-    it('should execute "eth_call" against mirror node with an undefined ETH_CALL_DEFAULT_TO_CONSENSUS_NODE', async function () {
-      const initialEthCallConesneusFF = process.env.ETH_CALL_DEFAULT_TO_CONSENSUS_NODE;
-
-      delete process.env.ETH_CALL_DEFAULT_TO_CONSENSUS_NODE;
-      restMock.onGet(`contracts/${defaultCallData.from}`).reply(404);
-      restMock.onGet(`accounts/${defaultCallData.from}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: defaultCallData.from,
-      });
-      restMock.onGet(`contracts/${defaultCallData.to}`).reply(200, defaultContract);
-      await ethImpl.call({ ...defaultCallData, gas: `0x${defaultCallData.gas.toString(16)}` }, 'latest');
-
-      assert(callMirrorNodeSpy.calledOnce);
-      process.env.ETH_CALL_DEFAULT_TO_CONSENSUS_NODE = initialEthCallConesneusFF;
-    });
-
-    it('should execute "eth_call" against mirror node with a ETH_CALL_DEFAULT_TO_CONSENSUS_NODE set to true', async function () {
-      const initialEthCallConesneusFF = process.env.ETH_CALL_DEFAULT_TO_CONSENSUS_NODE;
-
-      process.env.ETH_CALL_DEFAULT_TO_CONSENSUS_NODE = 'true';
-      restMock.onGet(`contracts/${defaultCallData.from}`).reply(404);
-      restMock.onGet(`accounts/${defaultCallData.from}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: defaultCallData.from,
-      });
-      restMock.onGet(`contracts/${defaultCallData.to}`).reply(200, defaultContract);
-      await ethImpl.call({ ...defaultCallData, gas: `0x${defaultCallData.gas.toString(16)}` }, 'latest');
-
-      assert(callConsensusNodeSpy.calledOnce);
-      process.env.ETH_CALL_DEFAULT_TO_CONSENSUS_NODE = initialEthCallConesneusFF;
-    });
-
-    it('gas exceeds limit', async function () {
-      restMock.onGet(`contracts/${accountAddress1}`).reply(404);
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-
-      const excessiveGasLimit = '15000001';
-      await ethCallFailing(
-        ethImpl,
-        {
-          from: accountAddress1,
-          to: contractAddress2,
-          data: contractCallData,
-          gas: excessiveGasLimit,
-        },
-        'latest',
-        (error) => {
-          expect(error).to.be.not.null;
-          expect(error.code).to.equal(predefined.GAS_LIMIT_TOO_HIGH(excessiveGasLimit, constants.BLOCK_GAS_LIMIT).code);
-        },
-      );
-    });
-
-    it('block 0', async function () {
-      restMock.onGet(`contracts/${accountAddress1}`).reply(404);
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-
-      const block = '0';
-      await ethCallFailing(
-        ethImpl,
-        {
-          from: accountAddress1,
-          to: contractAddress2,
-          data: contractCallData,
-          gas: maxGasLimitHex,
-        },
-        block,
-        (error) => {
-          const predefineError = predefined.UNSUPPORTED_HISTORICAL_EXECUTION(block);
-          expect(error.code).to.equal(predefineError.code);
-          expect(error.name).to.equal(predefineError.name);
-          expect(error.message).to.equal(predefineError.message);
-        },
-      );
-    });
-
-    it('block 1', async function () {
-      restMock.onGet(`contracts/${accountAddress1}`).reply(404);
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-
-      const block = '1';
-      await ethCallFailing(
-        ethImpl,
-        {
-          from: accountAddress1,
-          to: contractAddress2,
-          data: contractCallData,
-          gas: maxGasLimitHex,
-        },
-        block,
-        (error) => {
-          const predefineError = predefined.UNSUPPORTED_HISTORICAL_EXECUTION(block);
-          expect(error.code).to.equal(predefineError.code);
-          expect(error.name).to.equal(predefineError.name);
-          expect(error.message).to.equal(predefineError.message);
-        },
-      );
-    });
-
-    it('block earliest', async function () {
-      restMock.onGet(`contracts/${accountAddress1}`).reply(404);
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-
-      const block = 'earliest';
-      await ethCallFailing(
-        ethImpl,
-        {
-          from: accountAddress1,
-          to: contractAddress2,
-          data: contractCallData,
-          gas: maxGasLimitHex,
-        },
-        block,
-        (error) => {
-          const predefineError = predefined.UNSUPPORTED_HISTORICAL_EXECUTION(block);
-          expect(error.code).to.equal(predefineError.code);
-          expect(error.name).to.equal(predefineError.name);
-          expect(error.message).to.equal(predefineError.message);
-        },
-      );
-    });
-
-    it('block hash not supported', async function () {
-      restMock.onGet(`contracts/${accountAddress1}`).reply(404);
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-      restMock.onGet(`blocks?limit=1&order=desc`).reply(202);
-
-      await ethCallFailing(
-        ethImpl,
-        {
-          from: accountAddress1,
-          to: contractAddress2,
-          data: contractCallData,
-          gas: maxGasLimitHex,
-        },
-        blockHashTrimmed,
-        (error) => {
-          const predefineError = predefined.UNSUPPORTED_OPERATION(
-            `BlockParam: ${blockHashTrimmed} is not a supported eth_call block identifier`,
-          );
-          expect(error.code).to.equal(predefineError.code);
-          expect(error.name).to.equal(predefineError.name);
-          expect(error.message).to.equal(predefineError.message);
-        },
-      );
-    });
-
-    it('latest block but not found for comparison', async function () {
-      restMock.onGet(`contracts/${accountAddress1}`).reply(404);
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-      restMock.onGet(`blocks?limit=1&order=desc`).reply(404);
-
-      const block = '0x10';
-      await ethCallFailing(
-        ethImpl,
-        {
-          from: accountAddress1,
-          to: contractAddress2,
-          data: contractCallData,
-          gas: maxGasLimitHex,
-        },
-        block,
-        (error) => {
-          const predefineError = predefined.RESOURCE_NOT_FOUND(`unable to retrieve latest block from mirror node`);
-          expect(error.code).to.equal(predefineError.code);
-          expect(error.name).to.equal(predefineError.name);
-          expect(error.message).to.equal(predefineError.message);
-        },
-      );
-    });
-
-    it('to field is not a contract or token', async function () {
-      restMock.onGet(`contracts/${accountAddress1}`).reply(404);
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(404);
-      restMock.onGet(`tokens/${contractId2}`).reply(404);
-      web3Mock.onPost(`contracts/call`).reply(200, { result: '0x1' });
-
-      await expect(
-        ethImpl.call(
-          {
-            from: accountAddress1,
-            to: contractAddress2,
-            data: contractCallData,
-            gas: maxGasLimitHex,
-          },
-          'latest',
-        ),
-      ).to.eventually.be.fulfilled.and.equal('0x1');
-    });
-
-    // support for web3js.
-    it('the input is set with the encoded data for the data field', async function () {
-      restMock.onGet(`contracts/${accountAddress1}`).reply(200);
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200);
-      restMock.onGet(`tokens/${contractId2}`).reply(200);
-      web3Mock.onPost(`contracts/call`).reply(200, { result: '0x1' });
-
-      await expect(
-        ethImpl.call(
-          {
-            from: accountAddress1,
-            to: contractAddress2,
-            input: contractCallData,
-            gas: maxGasLimitHex,
-          },
-          'latest',
-        ),
-      ).to.eventually.be.fulfilled.and.equal('0x1');
-    });
-  });
-
-  describe('eth_call using consensus node', async function () {
-    let initialEthCallConesneusFF;
-
-    before(() => {
-      initialEthCallConesneusFF = process.env.ETH_CALL_DEFAULT_TO_CONSENSUS_NODE;
-      process.env.ETH_CALL_DEFAULT_TO_CONSENSUS_NODE = 'true';
-    });
-
-    after(() => {
-      process.env.ETH_CALL_DEFAULT_TO_CONSENSUS_NODE = initialEthCallConesneusFF;
-    });
-
-    it('eth_call with no gas', async function () {
-      restMock.onGet(`contracts/${accountAddress1}`).reply(404);
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-
-      sdkClientStub.submitContractCallQueryWithRetry.returns({
-        asBytes: function () {
-          return Uint8Array.of(0);
-        },
-      });
-
-      const result = await ethImpl.call(
-        {
-          from: accountAddress1,
-          to: contractAddress2,
-          data: contractCallData,
-        },
-        'latest',
-      );
-
-      sinon.assert.calledWith(
-        sdkClientStub.submitContractCallQueryWithRetry,
-        contractAddress2,
-        contractCallData,
-        400_000,
-        accountAddress1,
-        'eth_call',
-      );
-      expect(result).to.equal('0x00');
-    });
-
-    it('eth_call with no data', async function () {
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-      sdkClientStub.submitContractCallQueryWithRetry.returns({
-        asBytes: function () {
-          return Uint8Array.of(0);
-        },
-      });
-
-      const result = await ethImpl.call(
-        {
-          from: accountAddress1,
-          to: contractAddress2,
-          gas: maxGasLimitHex,
-        },
-        'latest',
-      );
-
-      sinon.assert.calledWith(
-        sdkClientStub.submitContractCallQueryWithRetry,
-        contractAddress2,
-        undefined,
-        maxGasLimit,
-        accountAddress1,
-        'eth_call',
-      );
-      expect(result).to.equal('0x00');
-    });
-
-    it('eth_call with no "from" address', async function () {
-      const callData = {
-        ...defaultCallData,
-        from: undefined,
-        to: contractAddress2,
-        data: contractCallData,
-        gas: maxGasLimit,
-      };
-
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-      sdkClientStub.submitContractCallQueryWithRetry.returns({
-        asBytes: function () {
-          return Uint8Array.of(0);
-        },
-      });
-
-      const result = await ethImpl.call(callData, 'latest');
-      expect(result).to.equal('0x00');
-    });
-
-    it('eth_call with wrong "from" address', async function () {
-      const callData = {
-        ...defaultCallData,
-        from: '0x0000000000000000000000000000000000000000',
-        to: contractAddress2,
-        data: contractCallData,
-        gas: maxGasLimit,
-      };
-
-      const result = await ethImpl.call(callData, 'latest');
-
-      expect(result.name).to.equal('Non Existing Account Address');
-      expect(result.code).to.equal(-32014);
-      expect(result.message).to.equal(`Non Existing Account Address: ${callData.from}. Expected an Account Address.`);
-    });
-
-    it('eth_call with all fields', async function () {
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-      sdkClientStub.submitContractCallQueryWithRetry.returns({
-        asBytes: function () {
-          return Uint8Array.of(0);
-        },
-      });
-
-      const result = await ethImpl.call(
-        {
-          from: accountAddress1,
-          to: contractAddress2,
-          data: contractCallData,
-          gas: maxGasLimitHex,
-        },
-        'latest',
-      );
-
-      sinon.assert.calledWith(
-        sdkClientStub.submitContractCallQueryWithRetry,
-        contractAddress2,
-        contractCallData,
-        maxGasLimit,
-        accountAddress1,
-        'eth_call',
-      );
-      expect(result).to.equal('0x00');
-    });
-
-    //Return once the value, then it's being fetched from cache. After the loop we reset the sdkClientStub, so that it returns nothing, if we get an error in the next request that means that the cache was cleared.
-    it('eth_call should cache the response for 200ms', async function () {
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-      sdkClientStub.submitContractCallQueryWithRetry.returns({
-        asBytes: function () {
-          return Uint8Array.of(0);
-        },
-      });
-
-      for (let index = 0; index < 3; index++) {
-        const result = await ethImpl.call(
-          {
-            from: accountAddress1,
-            to: contractAddress2,
-            data: contractCallData,
-            gas: maxGasLimitHex,
-          },
-          'latest',
-        );
-        expect(result).to.equal('0x00');
-        await new Promise((r) => setTimeout(r, 50));
-      }
-
-      await new Promise((r) => setTimeout(r, 200));
-
-      const expectedError = predefined.INVALID_CONTRACT_ADDRESS(contractAddress2);
-      sdkClientStub.submitContractCallQueryWithRetry.throws(expectedError);
-      const call: string | JsonRpcError = await ethImpl.call(
-        {
-          from: accountAddress1,
-          to: contractAddress2,
-          data: contractCallData,
-          gas: maxGasLimitHex,
-        },
-        'latest',
-      );
-
-      expect(call.code).to.equal(expectedError.code);
-      expect(call.name).to.equal(expectedError.name);
-      expect(call.message).to.equal(expectedError.message);
-    });
-
-    describe('with gas > 15_000_000', async function () {
-      it('eth_call throws gasLimit too high error when gas exceeds limit', async function () {
-        restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-
-        sdkClientStub.submitContractCallQueryWithRetry.returns(undefined);
-
-        const args = [
-          {
-            to: contractAddress2,
-            data: contractCallData,
-            gas: 50_000_000,
-          },
-          'latest',
-        ];
-
-        await RelayAssertions.assertRejection(
-          predefined.GAS_LIMIT_TOO_HIGH(50000000, constants.BLOCK_GAS_LIMIT),
-          ethImpl.call,
-          false,
-          ethImpl,
-          args,
-        );
-      });
-    });
-
-    it('SDK returns a precheck error', async function () {
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-      sdkClientStub.submitContractCallQueryWithRetry.throws(
-        predefined.CONTRACT_REVERT(defaultErrorMessageText, defaultErrorMessageHex),
-      );
-
-      const result = await ethImpl.call(
-        {
-          from: accountAddress1,
-          to: contractAddress2,
-          data: contractCallData,
-          gas: maxGasLimitHex,
-        },
-        'latest',
-      );
-
-      expect(result).to.exist;
-      expect(result.code).to.equal(-32008);
-      expect(result.name).to.equal('Contract revert executed');
-      expect(result.message).to.equal(`execution reverted: ${defaultErrorMessageText}`);
-      expect(result.data).to.equal(defaultErrorMessageHex);
-    });
-
-    it('eth_call with wrong `to` field', async function () {
-      const args = [
-        {
-          from: contractAddress1,
-          to: wrongContractAddress,
-          data: contractCallData,
-          gas: maxGasLimitHex,
-        },
-        'latest',
-      ];
-
-      await RelayAssertions.assertRejection(
-        predefined.INVALID_CONTRACT_ADDRESS(wrongContractAddress),
-        ethImpl.call,
-        false,
-        ethImpl,
-        args,
-      );
-    });
-
-    it('eth_call throws internal error when consensus node times out and submitContractCallQueryWithRetry returns undefined', async function () {
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-
-      sdkClientStub.submitContractCallQueryWithRetry.returns(undefined);
-
-      const result = await ethImpl.call(
-        {
-          to: contractAddress2,
-          data: contractCallData,
-          gas: 5_000_000,
-        },
-        'latest',
-      );
-
-      expect(result).to.exist;
-      expect(result.code).to.equal(-32603);
-      expect(result.name).to.equal('Internal error');
-      expect(result.message).to.equal(
-        'Error invoking RPC: Invalid contractCallResponse from consensus-node: undefined',
-      );
-    });
-  });
-
-  describe('eth_call using mirror node', async function () {
-    const defaultCallData = {
-      gas: 400000,
-      value: null,
-    };
-    let initialEthCallConesneusFF;
-
-    before(() => {
-      initialEthCallConesneusFF = process.env.ETH_CALL_DEFAULT_TO_CONSENSUS_NODE;
-      process.env.ETH_CALL_DEFAULT_TO_CONSENSUS_NODE = 'false';
-    });
-
-    after(() => {
-      process.env.ETH_CALL_DEFAULT_TO_CONSENSUS_NODE = initialEthCallConesneusFF;
-    });
-
-    //temporary workaround until precompiles are implemented in Mirror node evm module
-    beforeEach(() => {
-      restMock.onGet(`tokens/${defaultContractResults.results[1].contract_id}`).reply(404, null);
-      web3Mock.reset();
-    });
-
-    it('eth_call with all fields, but mirror-node returns empty response', async function () {
-      const callData = {
-        ...defaultCallData,
-        from: accountAddress1,
-        to: contractAddress2,
-        data: contractCallData,
-        gas: maxGasLimit,
-      };
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract3EmptyBytecode);
-      web3Mock.onPost(`contracts/call`).replyOnce(200, {});
-
-      const result = await ethImpl.call(callData, 'latest');
-      expect(result).to.equal('0x');
-    });
-
-    it('eth_call with no gas', async function () {
-      const callData = {
-        ...defaultCallData,
-        from: accountAddress1,
-        to: contractAddress2,
-        data: contractCallData,
-      };
-
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-      web3Mock.onPost('contracts/call', { ...callData, estimate: false }).reply(200, { result: `0x00` });
-      const result = await ethImpl.call(callData, 'latest');
-      expect(result).to.equal('0x00');
-    });
-
-    it('eth_call with no data', async function () {
-      const callData = {
-        ...defaultCallData,
-        from: accountAddress1,
-        to: contractAddress2,
-        gas: maxGasLimit,
-      };
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-      web3Mock.onPost('contracts/call', { ...callData, estimate: false }).reply(200, { result: `0x00` });
-
-      const result = await ethImpl.call(callData, 'latest');
-      expect(result).to.equal('0x00');
-    });
-
-    it('eth_call with no from address', async function () {
-      const callData = {
-        ...defaultCallData,
-        to: contractAddress2,
-        data: contractCallData,
-        gas: maxGasLimit,
-      };
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-      web3Mock.onPost('contracts/call', { ...callData, estimate: false }).reply(200, { result: `0x00` });
-      const result = await ethImpl.call(callData, 'latest');
-      expect(result).to.equal('0x00');
-    });
-
-    it('eth_call with all fields', async function () {
-      const callData = {
-        ...defaultCallData,
-        from: accountAddress1,
-        to: contractAddress2,
-        data: contractCallData,
-        gas: maxGasLimit,
-      };
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-      web3Mock.onPost('contracts/call', { ...callData, estimate: false }).reply(200, { result: `0x00` });
-      const result = await ethImpl.call(callData, 'latest');
-      expect(result).to.equal('0x00');
-    });
-
-    it('eth_call with all fields but mirrorNode throws 429', async function () {
-      const callData = {
-        ...defaultCallData,
-        from: accountAddress1,
-        to: contractAddress2,
-        data: contractCallData,
-        gas: maxGasLimit,
-      };
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-      web3Mock.onPost('contracts/call', { ...callData, estimate: false }).reply(429, mockData.tooManyRequests);
-      const result = await ethImpl.call(callData, 'latest');
-      expect(result).to.be.not.null;
-      expect(result.code).to.eq(-32605);
-      expect(result.name).to.eq('IP Rate limit exceeded');
-    });
-
-    it('eth_call with all fields but mirrorNode throws 400', async function () {
-      const callData = {
-        ...defaultCallData,
-        from: accountAddress1,
-        to: contractAddress2,
-        data: contractCallData,
-        gas: maxGasLimit,
-      };
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      web3Mock.onPost('contracts/call', { ...callData, estimate: false }).reply(400, mockData.contractReverted);
-      const result = await ethImpl.call(callData, 'latest');
-      expect(result).to.be.not.null;
-      expect(result.code).to.eq(-32008);
-      expect(result.name).to.eq('Contract revert executed');
-      expect(result.message).to.contain(mockData.contractReverted._status.messages[0].message);
-    });
-
-    it('eth_call with all fields, but mirror node throws NOT_SUPPORTED', async function () {
-      const callData = {
-        ...defaultCallData,
-        from: accountAddress1,
-        to: contractAddress2,
-        data: contractCallData,
-        gas: maxGasLimit,
-      };
-
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-      web3Mock.onPost('contracts/call', { ...callData, estimate: false }).reply(501, mockData.notSupported);
-
-      sdkClientStub.submitContractCallQueryWithRetry.returns({
-        asBytes: function () {
-          return Uint8Array.of(0);
-        },
-      });
-
-      const result = await ethImpl.call(callData, 'latest');
-
-      sinon.assert.calledWith(
-        sdkClientStub.submitContractCallQueryWithRetry,
-        contractAddress2,
-        contractCallData,
-        maxGasLimit,
-        accountAddress1,
-        'eth_call',
-      );
-      expect(result).to.equal('0x00');
-    });
-
-    it('eth_call with all fields, but mirror node throws CONTRACT_REVERTED', async function () {
-      const callData = {
-        ...defaultCallData,
-        from: accountAddress1,
-        to: contractAddress2,
-        data: contractCallData,
-        gas: maxGasLimit,
-      };
-
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-      web3Mock.onPost('contracts/call', { ...callData, estimate: false }).reply(400, mockData.contractReverted);
-      sinon.reset();
-      const result = await ethImpl.call(callData, 'latest');
-      sinon.assert.notCalled(sdkClientStub.submitContractCallQueryWithRetry);
-      expect(result).to.not.be.null;
-      expect(result.code).to.eq(-32008);
-      expect(result.name).to.eq('Contract revert executed');
-      expect(result.message).to.contain(mockData.contractReverted._status.messages[0].message);
-    });
-
-    it('SDK returns a precheck error', async function () {
-      const callData = {
-        ...defaultCallData,
-        from: accountAddress1,
-        to: contractAddress2,
-        data: contractCallData,
-        gas: maxGasLimit,
-      };
-
-      restMock.onGet(`accounts/${accountAddress1}${noTransactions}`).reply(200, {
-        account: '0.0.1723',
-        evm_address: accountAddress1,
-      });
-      restMock.onGet(`contracts/${contractAddress2}`).reply(200, defaultContract2);
-
-      web3Mock.onPost('contracts/call', { ...callData, estimate: false }).reply(400, {
-        _status: {
-          messages: [
-            {
-              message: '',
-              detail: defaultErrorMessageText,
-              data: defaultErrorMessageHex,
-            },
-          ],
-        },
-      });
-
-      const result = await ethImpl.call(callData, 'latest');
-
-      expect(result).to.exist;
-      expect(result.code).to.eq(-32008);
-      expect(result.name).to.eq('Contract revert executed');
-      expect(result.message).to.equal(`execution reverted: ${defaultErrorMessageText}`);
-      expect(result.data).to.equal(defaultErrorMessageHex);
-    });
-
-    it('eth_call with missing `to` field', async function () {
-      const args = [
-        {
-          ...defaultCallData,
-          from: contractAddress1,
-          data: contractCallData,
-          gas: maxGasLimit,
-        },
-        'latest',
-      ];
-
-      await RelayAssertions.assertRejection(
-        predefined.INVALID_CONTRACT_ADDRESS(undefined),
-        ethImpl.call,
-        false,
-        ethImpl,
-        args,
-      );
-    });
-
-    it('eth_call with wrong `to` field', async function () {
-      const args = [
-        {
-          ...defaultCallData,
-          from: contractAddress1,
-          data: contractCallData,
-          gas: maxGasLimit,
-        },
-        'latest',
-      ];
-
-      await RelayAssertions.assertRejection(
-        predefined.INVALID_CONTRACT_ADDRESS(wrongContractAddress),
-        ethImpl.call,
-        false,
-        ethImpl,
-        args,
-      );
-    });
-  });
-
-  describe('eth_sendRawTransaction', async function () {
-    const accountAddress = '0x9eaee9E66efdb91bfDcF516b034e001cc535EB57';
-    const accountEndpoint = `accounts/${accountAddress}${noTransactions}`;
-    const gasPrice = '0xad78ebc5ac620000';
-    const transactionIdServicesFormat = '0.0.902@1684375868.230217103';
-    const transactionId = '0.0.902-1684375868-230217103';
-    const value = '0x511617DE831B9E173';
-    const contractResultEndpoint = `contracts/results/${transactionId}`;
-    const ethereumHash = '0x6d20b034eecc8d455c4c040fb3763082d499353a8b7d318b1085ad8d7de15f7e';
-
-    this.beforeEach(() => {
-      sinon.restore();
-      sdkClientStub = sinon.createStubInstance(SDKClient);
-      sinon.stub(hapiServiceInstance, 'getSDKClient').returns(sdkClientStub);
-    });
-
-    this.afterEach(() => {
-      sinon.restore();
-    });
-
-    it('should return a predefined GAS_LIMIT_TOO_HIGH instead of NUMERIC_FAULT as precheck exception', async function () {
-      // tx with 'gasLimit: BigNumber { value: "30678687678687676876786786876876876000" }'
-      const txHash =
-        '0x02f881820128048459682f0086014fa0186f00901714801554cbe52dd95512bedddf68e09405fba803be258049a27b820088bab1cad205887185174876e80080c080a0cab3f53602000c9989be5787d0db637512acdd2ad187ce15ba83d10d9eae2571a07802515717a5a1c7d6fa7616183eb78307b4657d7462dbb9e9deca820dd28f62';
-      await RelayAssertions.assertRejection(
-        predefined.GAS_LIMIT_TOO_HIGH(null, null),
-        ethImpl.sendRawTransaction,
-        false,
-        ethImpl,
-        [txHash],
-      );
-    });
-
-    it('should return a computed hash if unable to retrieve EthereumHash from record due to contract revert', async function () {
-      restMock.onGet(accountEndpoint).reply(200, {
-        account: accountAddress,
-        balance: {
-          balance: Hbar.from(100_000_000_000, HbarUnit.Hbar).to(HbarUnit.Tinybar),
-        },
-      });
-
-      const transaction = {
-        chainId: 0x12a,
-        to: accountAddress1,
-        from: accountAddress,
-        value,
-        gasPrice,
-        gasLimit: maxGasLimitHex,
-      };
-
-      const signed = await signTransaction(transaction);
-
-      restMock.onGet(`transactions/${transactionId}`).reply(200, null);
-
-      const resultingHash = await ethImpl.sendRawTransaction(signed, getRequestId());
-      expect(resultingHash).to.equal(ethereumHash);
-    });
-
-    it('should throw internal error when transaction returned from mirror node is null', async function () {
-      restMock.onGet(accountEndpoint).reply(200, {
-        account: accountAddress,
-        balance: {
-          balance: Hbar.from(100_000_000_000, HbarUnit.Hbar).to(HbarUnit.Tinybar),
-        },
-      });
-
-      const transaction = {
-        chainId: 0x12a,
-        to: accountAddress1,
-        from: accountAddress,
-        value,
-        gasPrice,
-        gasLimit: maxGasLimitHex,
-      };
-
-      const signed = await signTransaction(transaction);
-
-      restMock.onGet(contractResultEndpoint).reply(404, mockData.notFound);
-      restMock.onGet(`transactions/${transactionId}?nonce=0`).reply(200, null);
-
-      sdkClientStub.submitEthereumTransaction.returns({
-        transactionId: TransactionId.fromString(transactionIdServicesFormat),
-      });
-
-      const response = (await ethImpl.sendRawTransaction(signed, getRequestId())) as JsonRpcError;
-
-      expect(response.code).to.equal(predefined.INTERNAL_ERROR().code);
-      expect(`Error invoking RPC: ${response.message}`).to.equal(predefined.INTERNAL_ERROR(response.message).message);
-    });
-
-    it('should throw internal error when transactionID is invalid', async function () {
-      restMock.onGet(accountEndpoint).reply(200, {
-        account: accountAddress,
-        balance: {
-          balance: Hbar.from(100_000_000_000, HbarUnit.Hbar).to(HbarUnit.Tinybar),
-        },
-      });
-
-      const transaction = {
-        chainId: 0x12a,
-        to: accountAddress1,
-        from: accountAddress,
-        value,
-        gasPrice,
-        gasLimit: maxGasLimitHex,
-      };
-
-      const signed = await signTransaction(transaction);
-
-      restMock.onGet(contractResultEndpoint).reply(200, { hash: ethereumHash });
-
-      sdkClientStub.submitEthereumTransaction.returns({
-        transactionId: '',
-      });
-
-      const response = (await ethImpl.sendRawTransaction(signed, getRequestId())) as JsonRpcError;
-
-      expect(response.code).to.equal(predefined.INTERNAL_ERROR().code);
-      expect(`Error invoking RPC: ${response.message}`).to.equal(predefined.INTERNAL_ERROR(response.message).message);
-    });
-
-    it('should return hash from ContractResult mirror node api', async function () {
-      restMock.onGet(accountEndpoint).reply(200, {
-        account: accountAddress,
-        balance: {
-          balance: Hbar.from(100_000_000_000, HbarUnit.Hbar).to(HbarUnit.Tinybar),
-        },
-      });
-
-      restMock.onGet(contractResultEndpoint).reply(200, { hash: ethereumHash });
-      const transaction = {
-        chainId: 0x12a,
-        to: accountAddress1,
-        from: accountAddress,
-        value,
-        gasPrice,
-        gasLimit: maxGasLimitHex,
-      };
-
-      sdkClientStub.submitEthereumTransaction.returns({
-        transactionId: TransactionId.fromString(transactionIdServicesFormat),
-      });
-      const signed = await signTransaction(transaction);
-
-      const resultingHash = await ethImpl.sendRawTransaction(signed, getRequestId());
-      expect(resultingHash).to.equal(ethereumHash);
-    });
   });
 
   describe('eth_getStorageAt', async function () {
